+++ src/crowdstream/cv/signal/matrix_ops.py
